pyproject.toml
[tool.poetry]
name = "cpython-python-api"
version = "0.1.0"
description = "CPython Python API"
authors = ["tuliomitico <tuliofc2008@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.0"
MarkupSafe = "~2.0.1"
Flask = "~1.1.4"
Flask-Classful = "^0.14.2"
python-dotenv = "^0.20.0"
flask-mongoengine = "^1.0.0"
Flask-JWT-Extended = "~4.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
blinker = "^1.4"
waitress = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "~6.0"
pytest-mock = "^3.8.1"
mongomock = "^4.0.0"
requests-mock = "^1.9.3"

[tool.pytest.ini_options]
markers = [
    # Define our new marker
    "unit: tests that are isolated from the db, external api calls and other mockable internal code."
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
